requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.25.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
datasets>=2.14.0

# Evaluation
sacrebleu>=2.3.1
nltk>=3.8

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
sentencepiece>=0.1.99

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0